# backend/docs/openapi/OverfitterAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "Overfitter",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "localhost:4000"
    },
    {
      "url": "overfitter.jamkelley.com/api"
    }
  ],
  "paths": {
...
    "/feet": {
      "get": {
        "summary": "Get All Feet Items",
        "operationId": "getAllFeetItems",
        "tags": [
          "Feet"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/FeetResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "post": {
        "summary": "Create Feet Item",
        "operationId": "createFeetItem",
        "tags": [
          "Feet"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/FeetResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete All Feet Items",
        "operationId": "deleteAllFeetItems",
        "tags": [
          "Feet"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/FeetResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/feet/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/idParam"
        }
      ],
      "get": {
        "summary": "Get Feet Item by Id",
        "operationId": "getFeetItemById",
        "tags": [
          "Feet"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/FeetResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete Feet Item by Id",
        "operationId": "deleteFeetById",
        "tags": [
          "Feet"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/FeetResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "put": {
        "summary": "Update Feet Item by Id",
        "operationId": "updateFeetItemById",
        "tags": [
          "Feet"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/FeetResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/feet/search": {
      "post": {
        "summary": "Search for Feet Item",
        "operationId": "searchFeetItem",
        "tags": [
          "Feet"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ClothingItemSearchBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/FeetResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/heads": {
      "get": {
        "summary": "Get All Head Items",
        "operationId": "getAllHeadItems",
        "tags": [
          "Heads"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HeadResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "post": {
        "summary": "Create Head Item",
        "operationId": "createHeadItem",
        "tags": [
          "Heads"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HeadResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete All Head Items",
        "operationId": "deleteAllHeadItems",
        "tags": [
          "Heads"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HeadResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/heads/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/idParam"
        }
      ],
      "get": {
        "summary": "Get Head Item by Id",
        "operationId": "getHeadItemById",
        "tags": [
          "Heads"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HeadResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete Head Item by Id",
        "operationId": "deleteHeadById",
        "tags": [
          "Heads"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HeadResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "put": {
        "summary": "Update Head Item by Id",
        "operationId": "updateHeadItemById",
        "tags": [
          "Heads"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HeadResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/heads/search": {
      "post": {
        "summary": "Search for Head Item",
        "operationId": "searchHeadItem",
        "tags": [
          "Heads"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ClothingItemSearchBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/HeadResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "FeetResponse": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/WithId"
                        },
                        {
                          "$ref": "#/components/schemas/Feet"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "FeetResponseList": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/WithId"
                          },
                          {
                            "$ref": "#/components/schemas/Feet"
                          }
                        ]
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "HeadResponse": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/WithId"
                        },
                        {
                          "$ref": "#/components/schemas/Head"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "HeadResponseList": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/WithId"
                          },
                          {
                            "$ref": "#/components/schemas/Head"
                          }
                        ]
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      },
      "UnexpctedError": {
        "description": "Unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      }
    },
    "schemas": {
      "WithId": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
...
      "DatalessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "item_status": {
        "type": "string",
        "enum": [
          "CLEAN",
          "DIRTY",
          "WORN",
          "OK",
          "UNKNOWN"
        ]
      },
      "item_condition": {
        "type": "string",
        "enum": [
          "NEW",
          "LIKE_NEW",
          "GOOD",
          "FAIR",
          "POOR",
          "TRASH",
          "UNKNOWN"
        ]
      },
      "size_us_letter": {
        "type": "string",
        "enum": [
          "5XS",
          "4XS",
          "3XS",
          "2XS",
          "XS",
          "S",
          "M",
          "L",
          "XL",
          "2XL",
          "3XL",
          "4XL",
          "5XL",
          "UNKNOWN"
        ]
      },
      "Entity": {
        "type": "object",
        "properties": {
          "uriImage": {
            "type": "string",
            "pattern": "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)",
            "example": "https://www.pngmart.com/files/11/Rickrolling-PNG-Transparent-Image.png"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "timestampAddedRFC": {
            "type": "string",
            "format": "date-time"
          },
          "timestampLastModifiedRFC": {
            "type": "string",
            "format": "date-time"
          },
          "rating": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5
          }
        }
      },
      "ClothingItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Entity"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "sizeUSLetter": {
                "$ref": "#/components/schemas/size_us_letter"
              },
              "sizeUSNumber": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "brand": {
                "type": "string"
              },
              "timestampPurchasedRFC": {
                "type": "string",
                "format": "date-time"
              },
              "itemCondition": {
                "$ref": "#/components/schemas/item_condition"
              },
              "itemStatus": {
                "$ref": "#/components/schemas/item_status"
              },
              "numberOfWears": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "wearsBeforeDirty": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "wearsLeftBeforeDirty": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "primaryColor": {
                "type": "string"
              },
              "secondaryColor": {
                "type": "string"
              },
              "accentColor": {
                "type": "string"
              },
              "pattern": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "feet_type": {
        "type": "string",
        "enum": [
          "DRESS_SHOE",
          "TENNIS_SHOE",
          "CASUAL_SHOE",
          "HEEL_SHOE",
          "FLIP_FLOP_SHOE",
          "SANDAL_SHOE",
          "DRESS_SOCK",
          "CASUAL_SOCK",
          "UNKNOWN"
        ]
      },
      "Feet": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ClothingItem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "$ref": "#/components/schemas/feet_type"
              },
              "sizeInches": {
                "type": "number",
                "minimum": 0,
                "maximum": 50
              }
            }
          }
        ]
      },
      "head_type": {
        "type": "string",
        "enum": [
          "BASEBALL_HAT",
          "BEANIE",
          "BANDANA",
          "GLASSES",
          "UNKNOWN"
        ]
      },
      "Head": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ClothingItem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "$ref": "#/components/schemas/head_type"
              },
              "sizeInchesLow": {
                "type": "number",
                "minimum": 0,
                "maximum": 50
              },
              "sizeInchesHigh": {
                "type": "number",
                "minimum": 0,
                "maximum": 50
              }
            }
          }
        ]
      },
...
      "ErrorType": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "errorMessage": {
            "type": "string"
          },
          "error": {
            "type": "object",
            "nullable": true
          },
          "detail": {
            "type": "object",
            "nullable": true
          }
        }
      }
    },
    "parameters": {
      "idParam": {
        "name": "id",
        "in": "path",
        "description": "ID of the Item",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "requestBodies": {
...
      "EntitySearchBody": {
        "description": "A JSON object containing search info for an Entity",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "timestampAddedBeginRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampAddedEndRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampModifiedBeginRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampModifiedEndRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "rating": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "ClothingItemSearchBody": {
        "description": "A JSON object containing search info for a ClothingItem",
        "content": {
          "application/json": {
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/requestBodies/EntitySearchBody"
                },
                {
                  "type": "object",
                  "properties": {
                    "sizeUSLetter": {
                      "type": "string"
                    },
                    "sizeUSNumber": {
                      "type": "integer"
                    },
                    "brand": {
                      "type": "string"
                    },
                    "timestampPurchasedBeginRFC": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "timestampPurchasedEndRFC": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "itemCondition": {
                      "type": "string"
                    },
                    "itemStatus": {
                      "type": "string"
                    },
                    "numberOfWearsMin": {
                      "type": "integer"
                    },
                    "numberOfWearsMax": {
                      "type": "integer"
                    },
                    "wearsBeforeDirtyMin": {
                      "type": "integer"
                    },
                    "wearsBeforeDirtyMax": {
                      "type": "integer"
                    },
                    "wearsLeftBeforeDirtyMin": {
                      "type": "integer"
                    },
                    "wearsLeftBeforeDirtyMax": {
                      "type": "integer"
                    },
                    "primaryColor": {
                      "type": "string"
                    },
                    "secondaryColor": {
                      "type": "string"
                    },
                    "accentColor": {
                      "type": "string"
                    },
                    "pattern": {
                      "type": "string"
                    },
                    "itemType": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}